Файлы подключены как зависимости CommonJS
