+++ src/form.js
@@ -2,7 +2,8 @@
 
 var browserCookies = require('browser-cookies');
 
-window.form = (function() {
+module.exports = (function() {
+
   var formContainer = document.querySelector('.overlay-container');
   var formCloseButton = document.querySelector('.review-form-close');
